throttler: Remove explicit cast from constants because it's not necessary.

+++ go/vt/throttler/throttler.go
@@ -32,7 +32,7 @@ const (
 
 	// MaxRateModuleDisabled can be set in NewThrottler() to disable throttling
 	// by a fixed rate.
-	MaxRateModuleDisabled = int64(math.MaxInt64)
+	MaxRateModuleDisabled = math.MaxInt64
 
 	// InvalidMaxRate is a constant which will fail in a NewThrottler() call.
 	// It should be used when returning maxRate in an error case.
@@ -40,7 +40,7 @@ const (
 
 	// ReplicationLagModuleDisabled can be set in NewThrottler() to disable
 	// throttling based on the MySQL replication lag.
-	ReplicationLagModuleDisabled = int64(math.MaxInt64)
+	ReplicationLagModuleDisabled = math.MaxInt64
 
 	// InvalidMaxReplicationLag is a constant which will fail in a NewThrottler()
 	// call. It should be used when returning maxReplicationlag in an error case.
